+++ github/stargazers/get.go
@@ -55,7 +55,7 @@ func processRepo(
 		stargazers = append(stargazers, star.User.GetLogin())
 	}
 	return model.Star{
-		RepoID:     int64(*repo.ID),
+		RepoID:     *repo.ID,
 		RepoName:   *repo.FullName,
 		Stargazers: stargazers,
 	}, nil
